# openapi-schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GitHub Trending API",
    "version": "1.0.0",
    "description": "Live GitHub repository data — trending AI repos, most starred, search, repo stats, contributors, releases. Uses official GitHub REST API."
  },
  "servers": [
    { "url": "https://serverless.on-demand.io/apps/github-trending" }
  ],
  "paths": {
    "/github/trending-ai": {
      "get": {
        "operationId": "getTrendingAI",
        "summary": "Get trending AI/ML/LLM repositories",
        "description": "Returns top AI repos pushed in the last N days, sorted by stars. Filters for LLM, AI agents, ML, generative AI topics with 100+ stars.",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 }, "description": "Max results (1-30)" },
          { "name": "days", "in": "query", "schema": { "type": "integer", "default": 30 }, "description": "Look back N days" }
        ],
        "responses": {
          "200": {
            "description": "Trending AI repos",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RepoListResponse" } } }
          }
        }
      }
    },
    "/github/trending": {
      "get": {
        "operationId": "getTrendingByTopic",
        "summary": "Get trending repos by topic",
        "description": "Returns repos created in the last N days for a specific topic, sorted by stars.",
        "parameters": [
          { "name": "topic", "in": "query", "schema": { "type": "string", "default": "ai" }, "description": "GitHub topic (e.g. ai, llm, machine-learning)" },
          { "name": "days", "in": "query", "schema": { "type": "integer", "default": 7 }, "description": "Look back N days" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "Trending repos",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RepoListResponse" } } }
          }
        }
      }
    },
    "/github/most-starred": {
      "get": {
        "operationId": "getMostStarred",
        "summary": "Get most starred AI repos all-time",
        "parameters": [
          { "name": "topic", "in": "query", "schema": { "type": "string", "default": "artificial-intelligence" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "Most starred repos",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RepoListResponse" } } }
          }
        }
      }
    },
    "/github/repo/{owner}/{repo}": {
      "get": {
        "operationId": "getRepo",
        "summary": "Get details for a specific repository",
        "parameters": [
          { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Repository details",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/github/repo/{owner}/{repo}/stats": {
      "get": {
        "operationId": "getRepoStats",
        "summary": "Get extended repo stats — contributors, releases, recent commits",
        "parameters": [
          { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Repository extended stats",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/github/search": {
      "get": {
        "operationId": "searchRepos",
        "summary": "Search GitHub repositories",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Search query (uses GitHub search syntax)" },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["stars", "forks", "updated", "help-wanted-issues"], "default": "stars" } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RepoListResponse" } } }
          }
        }
      }
    },
    "/github/user/{username}": {
      "get": {
        "operationId": "getUser",
        "summary": "Get GitHub user info and top repos",
        "parameters": [
          { "name": "username", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/github/health": {
      "get": {
        "operationId": "healthCheck",
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service healthy",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RepoListResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "total_count": { "type": "integer" },
          "count": { "type": "integer" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Repo" } }
        }
      },
      "Repo": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "full_name": { "type": "string" },
          "owner": { "type": "string" },
          "owner_avatar": { "type": "string" },
          "description": { "type": "string" },
          "url": { "type": "string" },
          "stars": { "type": "integer" },
          "forks": { "type": "integer" },
          "language": { "type": "string" },
          "topics": { "type": "array", "items": { "type": "string" } },
          "created_at": { "type": "string" },
          "updated_at": { "type": "string" }
        }
      }
    }
  }
}